use size_t

+++ sample/apps/blend/blend_sample.cpp
@@ -298,7 +298,7 @@ std::unique_ptr<RenderPipeline> BlendSample::createRenderPipeline(const BlendSta
         ShaderModuleDescriptor descriptor{};
         std::vector<char> vertexShaderSource = utils::readFile(m_appDir / "quad.vert.spv", m_handle);
         descriptor.type = ShaderModuleType::kSPIRV;
-        descriptor.code = std::string_view(vertexShaderSource.data(), static_cast<uint32_t>(vertexShaderSource.size()));
+        descriptor.code = std::string_view(vertexShaderSource.data(), vertexShaderSource.size());
 
         vertexShaderModule = m_device->createShaderModule(descriptor);
     }
@@ -339,7 +339,7 @@ std::unique_ptr<RenderPipeline> BlendSample::createRenderPipeline(const BlendSta
         ShaderModuleDescriptor descriptor{};
         std::vector<char> fragmentShaderSource = utils::readFile(m_appDir / "quad.frag.spv", m_handle);
         descriptor.type = ShaderModuleType::kSPIRV;
-        descriptor.code = std::string_view(fragmentShaderSource.data(), static_cast<uint32_t>(fragmentShaderSource.size()));
+        descriptor.code = std::string_view(fragmentShaderSource.data(), fragmentShaderSource.size());
 
         fragmentShaderModule = m_device->createShaderModule(descriptor);
     }
